fix the delete function and security configuration

+++ src/main/java/com/haidar/gestiondestock/service/Impl/CommandeFournisseurServiceImpl.java
@@ -130,6 +130,11 @@ public class CommandeFournisseurServiceImpl implements CommandeFournisseurServic
             log.error("Client is null");
             return;
         }
+        List<LigneCommandeFournisseur> ligneCommandeFournisseurs = ligneCommandeFournisseurRepository.findAllByCommandeFournisseurId(id);
+        if (!ligneCommandeFournisseurs.isEmpty()) {
+            throw new InvalidOperationException("Impossible de supprimer une commande fournisseur deja utilisee",
+                    ErrorCodes.COMMANDE_FOURNISSEUR_ALREADY_IN_USE);
+        }
         commandeFournisseurRepository.deleteById(id);
     }
